Replace the API link at the footer with Discord link
Task 11:

Removed the API Link: Since we will no longer expose the API endpoint
to the users, remove this link from the webui site. Need to add
authentication to protect the endpoint.

Add Discord Link: Generated a link to the #general channel of our
discord server. User should be able to join our discord server by
clicking on the link.