Change upload button to verify hash before uploading
Since the model that a user will upload is big, we need
to calculate the checksum and verify that we did not have
this model before uploading. The current implementation will
do exactly that. Do a checksum, and query the backend to see
if there is such a model, the model information will be preserved
in a csv file. If there is no such model, the web page will initiate
the uploading process.