Add a static folder and a route to serve static files
Add a static folder and move all static icons into this folder,
so the service will not expose html folder, where are templates of
html files, and could expose vanurebility of the service.

Add a `/public` route to serve static file, which is different than
the gradio one. The gradio app uses `/static` route to servce its static files.