Change button uppy binding from setTimeout to MutationObserver
We use to bind uppy button listener to gradio.HTML on change event,
however, this event is not reliable and we also have to set a timeout to
wait for dom ready. Nowm, we use a MutationObserver, we will be much
confident that it will observe the change and when it is ready, it will
always be able to bind the uppy button.